# perf: removed jsonOutput variable because is useless